% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVD.R
\name{newDImputation}
\alias{newDImputation}
\title{SVD imputation}
\usage{
newDImputation(dataX, Ncomp)
}
\arguments{
\item{dataX}{Missing data to be imputed}

\item{Ncomp}{the rank of the SVD approximation}
}
\value{

}
\description{
SVD imputation
}
\examples{
SVDimp <- SVDImputation(dataX = "Cornell",Ncomp = 5)
}
