% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.R
\name{newkNNImputation}
\alias{newkNNImputation}
\title{kNN imputation}
\usage{
newkNNImputation(dataX, k)
}
\arguments{
\item{dataX}{Missing data to be imputed}

\item{k}{number of neighbors}
}
\value{
imputed data with the kNN method
}
\description{
kNN imputation
}
\examples{
KNNimp <- KNNImputation(dataX = "Cornell",k = 5)
}
