% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MICE.R
\name{newMICEImputation}
\alias{newMICEImputation}
\title{SVD imputation}
\usage{
newMICEImputation(dataX, m)
}
\arguments{
\item{dataX}{Missing data to be imputed}

\item{m}{number of imputations}
}
\value{

}
\description{
SVD imputation
}
\examples{
MICEimp <- SMICEmputation(dataX = "Cornell",m = 5)
}
